Refactorings, documentation and database backup
This adds missing documentation for classes and methods of the following packages:
- app;
- controller;
- exceptions;
- factory;
- models.dto;
- models.filters;
- services;
- utils.

Some refactorings have occurred, mainly in services and controller classes. The abstract
class Services is now generic and defines basic methods (create, update, delete etc.) as
abstract, forcing their implementations.

This also adds a database backup (data only) as it was in 2024-10-14.